162: deletion bug (#173)
* #162: team deletion is ready

* #162: fix race deletion

* #162: add production condition in Program.cs

+++ car-racing-tournament-api/Controllers/RaceController.cs
@@ -79,6 +79,10 @@ namespace car_racing_tournament_api.Controllers
                 return BadRequest(_configuration["ErrorMessages:SeasonArchived"]);
             }
 
+            var resultDeleteResultsByRace = await _resultService.DeleteResultsByRaceId(resultGet.Race.Id);
+            if (!resultDeleteResultsByRace.IsSuccess)
+                return NotFound(resultDeleteResultsByRace.ErrorMessage);
+
             var result = await _raceService.DeleteRace(resultGet.Race);
             if (!result.IsSuccess)
                 return BadRequest(result.ErrorMessage);
